requirements.txt
# LLM & Orchestration
openai>=1.40.0
groq>=0.11.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
tiktoken>=0.7.0

# Ingestion & Document Processing
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pytesseract>=0.3.13
Pillow>=10.4.0
tabula-py>=2.9.0
pandas>=2.2.0
numpy>=1.26.0

# Vector Store & Search
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0

# API & Auth
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1  # passlib not yet compatible with bcrypt 5.x
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
httpx>=0.27.0
tenacity>=8.5.0
rich>=13.7.0
